Change period of yearly temperature forcing to be 4 units of time.
At one point changed this to 8 to debug if top liquid region becoming
disconnected was causing the instability but it didn't appear to be the
problem. However physically if top region is disconnected from the ocean
by an impermeable section of ice and gas fluxes to the atmosphere in
theory liquid must come from somewhere to offset this loss of volume or
the ice desaturates.