add first attempt at lax friedrich solver
Seems to be completely unstable if we use lax friedrich for enthalpy so
just stick to using normal upwind method here. For the salt and gas I
think the boundary values are badly affecting the results and the fact
that we still have diffusion in the solid phase.

To deal with this we could write the method in flux form instead and use
the lax friedrich fluxes only when the medium is permeable.