Move phase boundary calculation into a class structure.
The abstract class PhaseBoundaries defines the interface. I.e. that a
phase boundary class should provide a method to calculate the masks
determining what phase each cell is in for a given state.

The concrete implementations are responsible for calculating the
phase boundaries for a given model. I have implemented one for the full
enthalpy method where thermal properties depend on gas fraction.