Get rid of unused numerical parameters and log if courant number.
Add an optional method to solver template which can be implemented to
run any checks on the paramters before simulation is run. In the lagged
upwind solver for full model I implement this to check if the Courant number
for thermal diffusion is greater than 0.5 which is the stability threshold
for explicit methods.