Add barrow surface temperature forcing read from 2009 data.
Put the source for the data and the metadata explaining how to read it
in celestine/forcing_data.

In order to read the data we must be able to convert from simulation
time to dimesnional time. This means to use the barrow temperature
forcing the simulation config must be created by the DimensionalParams
class. This will attach a scales object to the config which is otherwise
left as None. This is used to convert the simulation to real world time
which can then be interpolated in the air temperature data from the
barrow site. This forcing option is called "barrow_2009". I have done an
example using this in the barrow.py script which runs for about the
length of the data available (approximately 164 days) from the 15th
January 2009.