Move initial conditions calculation to separate module.
The interface is very similar to forcing. You provide a string choice in
the configuration boundary conditions of the initial condition you want.
Have renamed the default I was using to "uniform". The solver template
then implements the initial conditions by using this key to get the
correct function from the initial_conditions module.

This is necessary to implement a different initialisation for barrow
2009.