Refactor so that forcing config loads barrow data only if needed.
Had to do this as before it was being loaded on import of forcing.py
which was messing up the sphinx documentation.

Now forcing configuration has class variables with the barrow forcing
data path and indices hard coded in. It also has a function to load this
data and populate attributes to store the air temperature and time in
days. This is then used by the barrow forcing.

To make sure this is only loaded when necessary this method of forcing
config is called by the solver when initialising a simulation only if
the forcing choice is "barrow_2009"