Remove unecessary functions to calculate liquid Darcy velocity.
We no longer attempt to solve for liquid darcy velocity from pressure.
Instead we will parameterise the liquid darcy velocity and use this to
inform the gas velocity. This has yet to be implemented and so for now
the user provides the liquid interstitial velocity in the darcy law
params.

Removed functions from celestine.velocities to solve pressure equation
and to calculate permeability. This means changing the signature of the
exported function calculate_velocities that the solvers use.

This change DEPRECATES the lagged solver as this is the only solver to
attempt to calculate liquid darcy velocity from pressure field.