Change calculation of liquid darcy velocity to use brin convection.
Default of brine convection flag in config object is false. In which
case liquid velocity is just set to be zero everywhere in the domain.

However if it is set to true then the liquid velocity will be calculated
using the new brine_drainage module.

Note this new calculation hasn't been coupled to the gas velocity yet.