!! Remove data_path as a configuration parameter.
I want to be able to specify where configurations and simulation data is
saved when running the program therefore the data_path cannot be
specified in the configurations.

This will break scripts that currently produce and save configuration
files.
It will also break the functions that call solvers as these currently
use the data_path to specify the path to save the simulation data.

I have made the save functions for the DimensionalParams, Configuration
and solution classes take an extra argument to specify the directory to
save the data to.