Set up argparse in celestine/__main__.py.
This means the program can be run as python -m celestine --options.

Implemented options for directory to search for yaml configurations,
directory to save the output of the simulation to and a flag for if the
configurations are dimensional or not.

The program will glob all the yaml files in the configuration directory.
Then I still need to implement loading them all as configuration
objects, running the simulations and saving the output in the target
directory.