Use different thermal conducitvities for liquid and solid phase.
Change the dimensional parameter thermal_conductivity ->
liquid_thermal_conductivity. This will still be used to calculate all
the non-dimensional scales and diffusivity and will be the total
conductivity in the case where we don't phase average the conductivity
term as is currently inplemented.

However going to implement phase averaged thermal conductivity for solid
and liquid. In this case we will need both parameters.