Fix numerical instability with phase average heat conduction.
The max timestep parameter in scipy solver was calculated previously to
ensure timestep below the Courant stability condition for thermal
diffusion with a non-dimensional diffusivity of 1. Now that there is
enhanced heat transport in the solid phase this needs to be reduced
suitably for stability. The lower value of 0.1 should suffice for sea
ice values.