Add tolerable super saturation fraction parameter.
Add this to dimensional params and it gets carried over to
Config.PhysicalParams object.

This returns to normal case when we set it equal to 1

However we may now specify that brines can tolerate some supersaturation
before nucleation. No nucleation case would be setting this parameter to
be very large. This lets us approximately investigate incomplete
nucleation and so ice can reject some super saturated water to the ocean
if we assume the ocean is saturated.

Chaged the implementation of ReducedEnthalpyMethod to get gas fraction
and dissolved gas content as this is currently the only enthalpy method
we are using.