Refactor barrow.py script to use a main function.
The script generates simulation configuration from non-dimensional
parameters with barrow forcing and initial conditions. Then it runs the
simulation, saves the configurations and plots snapshots of the
solutions (frames) for key quantities.

Refactor this behaviour into a main function that accepts dimensional
simulation parameters and a directory for the configurations and
simulation data and a separate directory for the frame images to go.

This will allow us to run this script as a test for a variety of
different simulations.