Test if example script runs as intended.
Add pytest as a dependency to do this.

test_example.py has three integration tests for the example script in
celestine. It checks that generated configuration files are the same as
reference versions held in tests/reference_data. It then runs the script
with the parameters given in the script to check it doesn't raise any
errors.

Register a marker for slow tests in pytest config file.