Replace generate tests script with pytest function that runs them.
This old script generated 12 test cases that we then had to run.
Now pytest will use the yaml configurations stored in
test/configrations/YearlyConstant and call a subprocess to run the
command line interface to run all these simulations. Test passes if this
completes without errors and the data will be stored in a directory
created at the project root.