Refactor so that ode_fun in solver only depends on state class.
This means it is now independent on the form of the solution we are
assuming, i.e what components the vector has and so can be extended
without modification to different implementations of the state class.

Achieved this by adding a class method to the state class which
initialises it from the solution vector stacked form.