Create load module to load data produced by different models.
Export useful functions through the pacakge __init__.py.
Created functions to read in the data produced by different models and
produce the state at a specified time or produce an array of enthalpy
method quantities like temperature which can be used to make contour
plots.

This is also now illustrated in the example script.