Add summer initial condition.
This initialises the simulation with a layer of ice of fixed bulk
salinity, temperature and no gas.

This can be used along with constant temperature forcings as shown in
the new script added to diagnostics module.

Added the requuired dimensional parameters for initial ice salinity and
temperature and ocean temperature as well as their
non- dimensionalisation.