!! Make EQMStateBCs a dataclass and extract applying boundary conditions
Make EQMStateBCs just contain the state data after enthalpy method is
run wth boundary conditions added to the ghost grid. This means the
logic of applying the boundary conditions is moved to the
apply_boundary_conditions function.

This is a breaking change so will need more refactoring throughout the
code