Refactor calculating ode equation to own file.
While this is much cleaner it has had a terrible impact on performance
in the test cases (around 4x slower).

This is probable because we have to check each timestep what the type of
the model is even though this only really needs to be checked once
initially.

To get around this commented out in the equations file is an idea to
instead return a function that implements the required equations based
on the configuration