Add verbosity option to CLI of the form --verbose or -vv etc.
Define an optprint function in the ode solver functions and the main
entry point which takes an optional verbosity level and prints messages
only if this is greater than or equal to 1.

So program prints nothing if no verbosity flag is given on the command
line.

Program prints all messages if any of --verbose -v -vv etc are given.