Refactor conversions to dimensionless parameters.
Each dimensionless parameter module now handles converting the
dimensional parameter object to the dimensionless counterpart. This is
then combined in the configuration object.

This means where configuration needs no dimensional parameters such as
in the case of the NumericalParams or some initial conditions these
should be moved to the dimensional parameters module and just imported.