Do not use default values for non obvious configuraition options.
Removed most of the default values from Config and DimensionalParams
objects as depedning on the simulation this can cause confusion of
forgetting to change the initial conditions if you change the forcing
for example. Or forgetting to change the simulated length of time.

Left defaults only for values where this makes sense across simulations
such as gravity and the frame velocity as this will usually be zero.