Add escape_ice_surface bool param for bubbles.
This determines if gas / oil droplets will leave the top surface of the
domain. For simulating gas this is what we want as gas bubbles should
flux to the atmosphere. However, for simulating oil this is not what we
want as oil droplets should collect at the ice surface. This is
implemented in the function in the equations module to prevent
unphysical buoyant velocities in the gas phase using this parameter to
determine which case we are in.