Refactor temperature forcing to take full state.
This means it can solve for the appropriate ghost cell temperature to
impose using the non-linear surface energy balance.

This means we don't have to change the implementation based on a check
if we are doing radiative forcing or not.

Move the surface energy balance and calculation of turbulent heat fluxes
to their own submodule within the forcing module.