Break up RadForcing configuration options.
Create a short wave forcing configuration object to be composed with
RadForcing. This provides the parameters necessary to calculate incoming
dimensional shortwave radiative flux and we have implemented ConstantSW
to provide a constant value.

Seperate the parameters to calculate internal heating due to oil in the
simulation into the oil_heating option of RadForcing. We create three
different cases:

- NoHeating: where oil has no radiative effect
- BackgroundOilHeating: where internal heating is calculated using a
  constant backgroun concenctration of oil during the simulation using a
  single layer model from the package oilrad.
- MobileOilHeating: internal heating is calculated dynamically during
  the simulation based on the diagnosed vertical profile of oil mass
  fraction and passed to the solver in the package oilrad.

Finally we refactor the equations function to remove the check for if
the simulation is using radiative forcing and move this to the radiative
heating function which will just return an array of zeros in the case of
no radiative forcing.