Remove SW_albedo forcing option and infer droplet radius.
We no longer need to implement the shortwave forcing in the boundary
condition as it is resolved by the heating term. The albedo is to be
diagnosed from the simulation and not imposed. The
SW_penetration_fraction parameter now represents the fraction of
shortwave radiation that is not scattered away from the ice by the surface
scattering layer and so can contribute to internal heating.

Currently we do not allow for heating in the SSL but this could be put
back into the boundary condition if needed.

Also make the simulation get the droplet radius for the radiative
transfer calculation from the gas bubble radius specified in the mono
distributed case.