Edit radiative heating to use liquid fraction in domain.
We use only the infinite layer model from oilrad as this allows for
variable scattering in the domain.

Where there is no ice the scattering is set to zero by oilrad v0.5.0.

We also set the surface SW penetration fraction to 1 once a meltpond
forms as there can be no surface scattering layer in this case.