Add option to choose solver method in scipy solve_IVP.
Default is set to RK23 which is an explicit method and when running
explicit methods the timestep is calculated to be less than the courant
number for numerical stability.

We now have the option to choose an implicit method like Radau
which does not require a timestep restriction. This is important as it
avoids the very heavy timestep restriction of having larger thermal
conductivities in turbulent regions.