For ERA5 forcing with snow data estimate surface heat flux thtough snow
Using the given snow depth data assume a homogenous snow layer where we
just model steady sate conduction and use this to infer the heat flux
reaching the ice surface. This requires snow density to convert
equivalent water snow depth data which is hard coded in to the ERA5
forcing class. We also need the snow conductivity which is added. As the
snow depth goes below the threshold value the surface enery balance
condition returns to the bare ice surface version.