Make eddy diffusivity turn on smoothly in liquid using exponential.
exponential switch is scaled so that for liquid fraction less than 0.9
up to eddy diffusivity of 0.15m2/s (value from McPhee paper on oceanic
turbulence under ice) eddy diffusivity remains at least an order of
magnitude smaller than the thermal diffusivity. This is important so
that we do not accidentally increase diffusivity within the ice.