Upgrade oilrad 0.7.0 -> 0.9.9 for 2 stream shortwave radiative transfer
This new version of oilrad takes the ice_scattering_coefficient
appropriate for the two stream model directly. The default value is
derived from RedmondRoche2022 using the backscatter fraction for the
henyey-greenstein phase function and ice scattering cross section.

We also can optionally supply the absorption_enhancement_factor which is
an ad-hoc multiplier for the absorption cross section from
RedmondRoche2022 which accounts for the mean path length extension of
the diffuse scattered irradiance field. This is impractical to relate to
the phase function but the default value is found by fitting the albedo
profiles for 0.8m thick first year ice with no oil in RedmondRoche2022.