Update dependencies
Firstl we upgrade to oilrad v1.0.0 which requires updating to numpy
version 2. This requires changing all np.NaN to np.nan but apart from
that no further changes needed.

Also upgraded xarray and netCDF4.

Remove the group that matplotlib was in and make it a full dependency as
it is used by the seaice3p.plot module.

Remove unecessary dependencies snakviz and pydeps