Add advective_heat_flux_in_ocean boolean option to RJW14 parameters.
Default is true which leaves the constant upwelling veclocity in the
liquid coupled to the advetive flux of heat in the domain.

If this option is set to false then a tanh is used to set the convective
heat flux in the ocean to zero. The rationale is that this might be
swamped by turbulent motions in reality.

By having this option we can explore the consequences of both cases on
ice growth