Change to using the SixBandModel for SW radiation from oilrad 2.0.0
This provides a better description of transmission of visible
wavelengths through a surface snow layer and SSL.

Remove redundant configuration parameters for wavelength resolution and
the SSL penetration fraction as now transmission through the SSL is
computed in each wavelength band.

Add parameters specifying the snow and SSL albedo and extinction
coefficients in each wavelength band as well as the SSL thickness.

Remove the oil_simulation module which was not being used.

Remove calculation of PAR from the results class as there are methods to
do this in oilrad 2.0.0