FIX: Mobile oil simulation crash as dividing by zero liquid fraction
In some years the ice gets cold enough (sub-eutectic) for some cells to
have liquid_fraction==0.

In this case when calculating the brine channel sink for bubbles / oil
droplets we were dividing by zero causing a crash.

The fix was to check if liquid fraction is zero and replace the sink
term with zero where it is as no brine channels occur here anyway.